requirements.txt
fastapi
pytest
uvicorn
httpx
python-dotenv
websockets

kokoro
ipython
soundfile
torch

useful-moonshine-onnx@git+https://git@github.com/usefulsensors/moonshine.git#subdirectory=moonshine-onnx
